requirements-cpu.txt
# 使用 PyTorch 官方 CPU 源
--index-url https://download.pytorch.org/whl/cpu

torch==2.4.0
torchvision==0.19.0

einops>=0.7.0
numpy>=1.24
tqdm>=4.66
matplotlib>=3.8
pillow>=10.0
scikit-learn>=1.3

# 可选
# vit-pytorch==1.10.1
